发布考试

+++ src/main/java/com/czy/template/view/dto/PublicExamDTO.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 @Data
 public class PublicExamDTO {
-    private int examId;                   //考试id
+    private Long examId;                   //考试id
     private String title;                 //考试名称
     private List<Integer> questionIds;    //考试题目
     private int totalScore;               //总分数
